refactor(cell): shared to register loading using ldmatrix.
